+++ visualization/app/codeCharta/features/fileExtensionBar/components/fileExtensionBarSegment/fileExtensionBarSegment.component.spec.ts
@@ -38,6 +38,7 @@ describe("FileExtensionBarSegment", () => {
         preferences: {
             isPresentationMode: false,
             resetCameraIfNewFileIsLoaded: true,
+            centerMapZoom: 140,
             maxTreeMapFiles: 100,
             experimentalFeaturesEnabled: false,
             screenshotToClipboardEnabled: false,