+++ visualization/app/codeCharta/features/fileExtensionBar/services/blackListExtension.service.spec.ts
@@ -91,6 +91,7 @@ describe("BlackListExtensionService", () => {
         preferences: {
             isPresentationMode: false,
             resetCameraIfNewFileIsLoaded: true,
+            centerMapZoom: 140,
             maxTreeMapFiles: 100,
             experimentalFeaturesEnabled: false,
             screenshotToClipboardEnabled: false,