+++ visualization/app/codeCharta/stores/preferences/store/centerMapZoom/centerMapZoom.actions.ts
@@ -1,0 +1,3 @@
+import { createAction, props } from "@ngrx/store"
+
+export const setCenterMapZoom = createAction("SET_CENTER_MAP_ZOOM", props<{ value: number }>())